+++ Pieklo dzidowca kregi spierdolenia/Assets/Resources/Dialogues/Events/GetQuest.asset
@@ -9,8 +9,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bd5e6159b4d06d94791100962b751234, type: 3}
+  m_Script: {fileID: 11500000, guid: e4b2733d05b0f104187d13b3a7222286, type: 3}
   m_Name: GetQuest
   m_EditorClassIdentifier: 
-  statCheckType: 2
-  statCheckValue: 2
+  questName: "Walka o wolno\u015B\u0107"
