+++ Pieklo dzidowca kregi spierdolenia/Assets/Resources/Dialogues/Events/GetQuest.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 91c3d8002f390224ea5253a573fa4a67
+guid: 94ef35f1d6368c24a82953d77de19e1b
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 0
