--- Pieklo dzidowca kregi spierdolenia/Assets/Resources/Dialogues/Events/StatCheck.asset
@@ -1,16 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bd5e6159b4d06d94791100962b751234, type: 3}
-  m_Name: StatCheck
-  m_EditorClassIdentifier: 
-  statCheckType: 2
-  statCheckValue: 2
