+++ Pieklo dzidowca kregi spierdolenia/Assets/Scenes/StatsCreator.unity
@@ -1157,7 +1157,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b374279be5765df4599cd3d016916590, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  dialogueContainer: {fileID: 11400000, guid: fa76cf32b8be6494bb57b40917551cb4, type: 2}
+  dialogueContainer: {fileID: 0}
   dialogueController: {fileID: 1999005679}
   audioSource: {fileID: 501850837}
 --- !u!1 &505183178 stripped
