+++ Pieklo dzidowca kregi spierdolenia/Assets/Scripts/Dialogues/Dialogue Use/Events/EventGetQuest.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: bd5e6159b4d06d94791100962b751234
+guid: e4b2733d05b0f104187d13b3a7222286
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
