+++ Pieklo dzidowca kregi spierdolenia/Assets/Scripts/Items/Enums.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 3ddde5667ce13d7448dc6a4fd1b62897
-NativeFormatImporter:
+guid: 2ab8c8b45bdfbc34198250457ffe8dbb
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
