+++ Pieklo dzidowca kregi spierdolenia/Assets/Scripts/Items/Enums/EquipmentSlot.cs
@@ -1,0 +1,12 @@
+﻿namespace jbzdy.Items.Enums
+{
+    public enum EquipmentSlot
+    {
+        Head,
+        Chest,
+        Legs,
+        Feet,
+        Weapon,
+        Shield,
+    }
+}
