+++ Pieklo dzidowca kregi spierdolenia/Assets/Scripts/Items/Enums/EquipmentSlot.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f7a45889e10680347939c277a7904dc4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
