+++ Pieklo dzidowca kregi spierdolenia/Assets/Scripts/Items/ItemBase/EquipableItem.cs
@@ -1,36 +1,31 @@
-﻿using UnityEngine;
+﻿using jbzdy.Items.Enums;
+using UnityEngine;
 
 /// <summary>
 /// Napisane przez sharashino 
 /// 
 /// Przedmioty które możemy ubrać
 /// </summary>
-[CreateAssetMenu(fileName = "NewEquipableItem", menuName = "Items/New Equipable Item")]
-public class EquipableItem : Item
+namespace jbzdy.Items
 {
-    public EquipmentSlot equipmentSlot;
-    public int armorModifier;
-    public int damageModifier;
-
-    public override void Use()
+    [CreateAssetMenu(fileName = "NewEquipableItem", menuName = "Items/New Equipable Item")]
+    public class EquipableItem : Item
     {
-        base.Use();
+        public EquipmentSlot equipmentSlot;
+        public int armorModifier;
+        public int damageModifier;
 
-        //Equip item
-        //
-        //
-        //Remove from inventory
-        EquipmentManager.instance.EquipItem(this);
-        RemoveFromInventory();
+        public override void Use()
+        {
+            base.Use();
+
+            //Equip item
+            //
+            //
+            //Remove from inventory
+            EquipmentManager.instance.EquipItem(this);
+            RemoveFromInventory();
+        }
     }
 }
 
-public enum EquipmentSlot
-{
-    Head,
-    Chest,
-    Legs,
-    Feet,
-    Weapon,
-    Shield,
-}
