+++ Pieklo dzidowca kregi spierdolenia/Assets/Scripts/Stats/Base/CharacterStats.cs
@@ -1,7 +1,5 @@
 ﻿using UnityEngine;
-using System.Collections.Generic;
-using System;
-
+using jbzdy.CharacterStats.Stats;
 /// <summary>
 /// Napisane przez Sharashino
 ///
